moves ProjectSelect
adds some props
