+++ frontend/src/pages/projects/suite.js
@@ -6,6 +6,9 @@ const suite = create((data = {}, fieldName) => {
     test("nickname", "This is required information", () => {
         enforce(data.nickname).isNotBlank();
     });
+    test("nickname", "Nickname must be 3 letters", () => {
+        enforce(data.nickname).lengthEquals(3);
+    });
     test("title", "This is required information", () => {
         enforce(data.title).isNotBlank();
     });
